Fix seg fault in MonteCarloAbsorption
MonteCarloAbsorption would occasionally crash when running on Ubuntu
when a material on one of the shapes wasn't defined. In this situation
the code in one of the threads would create the missing material. Sometimes
multiple threads would create the missing material and it appears this
wasn't thread safe.
I have changed CSGObject::material so it doesn't create a dummy material if
it is missing